Separate TabPanel component
